[discretizers.builtin.grids] remove defaultimpl
- moves code from defaultimpl to interfaces
- rewraps docstrings
